# cia-backend/cia-partner-api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CIA Partner API",
    "description": "Open API for Insurtech connectivity — aggregators, digital brokers, and embedded insurance providers.",
    "version": "1.0.0",
    "contact": {
      "name": "CIA Partner Support",
      "email": "partners@cia.app"
    }
  },
  "servers": [
    { "url": "{{baseUrl}}/partner/v1", "description": "Partner API base" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/customers/individual": {
      "post": {
        "summary": "Register individual customer",
        "operationId": "registerIndividualCustomer",
        "tags": ["Customers"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "201": { "description": "Customer registered" }
        }
      }
    },
    "/customers/corporate": {
      "post": {
        "summary": "Register corporate customer",
        "operationId": "registerCorporateCustomer",
        "tags": ["Customers"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "201": { "description": "Customer registered" }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "summary": "Get customer",
        "operationId": "getCustomer",
        "tags": ["Customers"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Customer details" },
          "404": { "description": "Not found" }
        }
      }
    },
...
    "/claims/{id}": {
      "get": {
        "summary": "Get claim status",
        "operationId": "getClaim",
        "tags": ["Claims"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Claim details" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}